llm/conversational.py
import os
from langchain.prompts import PromptTemplate
from langchain_openai import ChatOpenAI
from langchain_groq import ChatGroq
from dotenv import load_dotenv



from calculations.cost_eval import compute_llm_cost
load_dotenv()

def _format_chunk_id_context(top_texts, chunk_ids):
    chunk_context = []
    for i, (text, chunk_id) in enumerate(zip(top_texts, chunk_ids)):
        snippet = text.strip()
        if not snippet:
            continue
        identifier = f"[Chunk {chunk_id}]" if chunk_id is not None else f"[{i+1}]"
        chunk_context.append(f"{identifier} {snippet}")
    return "\n\n".join(chunk_context)

def get_conversational_answer(top_texts, query, model_type="openai", past_turns_context="", chunk_ids=None):
    if chunk_ids is None:
        chunk_ids = [None] * len(top_texts)
    chunk_context = _format_chunk_id_context(top_texts, chunk_ids)

    template = """
SYSTEM INSTRUCTIONS (READ FIRST):
You are a friendly, concise, and factual bank assistant. Follow these rules exactly:

1) Use ONLY the information provided in the CONTEXT, CONVERSATION SO FAR blocks. 
   - Do NOT invent or assume information
   - If the answer is not present, respond exactly: 
     I apologize, I do not have knowledge about this

2) **For EMI calculations (Speacial rules)
    - REQUIRED INPUTS
	    - Loan Amount (Principal, in INR)
	    - Annual Interest Rate (in %)
 	    - Loan Tenure (in years)

    - INTEREST RATE PRIORITY    
   	    - First, check the CONTEXT and CONVERSATION SO FAR for the interest rate relevant to the users loan type.
	    - If the rate is found, use it directly.
	    - Only ask the user for the interest rate if its missing in both.
	- If LOAN AMOUNT or TENURE is missing, politely prompt the user to provide them. Do not proceed until all three values are available.
    - Use formula: EMI = P * r * (1+r)^n / ((1+r)^n-1) 
        where:
	        - P = Loan Amount
	        - r = Monthly Interest Rate = Annual Rate ÷ 12 ÷ 100
	        - n = Tenure in months (Years * 12)
   
   **IMPORTANT EMI RESPONSE FORMAT:**
    - Once all inputs are available, validate them by restating clearly.
    - Calculate and display:
		- Monthly EMI
		- Total repayment (EMI * n)
		- Total interest payable (Total repayment * Loan Amount)
   
   - EMI RESPONSE FORMAT (Mandatory)
        Loan Details (as per your input):
        - Loan Amount: ₹<amount>
        - Interest Rate: <rate>% per annum
        - Tenure: <years> years (<months> months)

        Result:
        - Monthly EMI: ₹<emi>
        - Total Repayment: ₹<total_repayment>
        - Total Interest: ₹<total_interest>

   - Do NOT show calculation steps, formulas, or mathematical workings. Only provide the final conversational result with chunk references.

   

3) Answering style:
   	- DIRECT /FACTUAL QUESTIONS (e.g., “What is the processing fee?”):
        - Answer in 2–3 short, conversational sentences.
        - Start with a light exclamation or emoji if appropriate, e.g., "Ahh! 😊", "Oh, I see! 👀", "Haha, got it! 😄", "Alright, let’s check this out! 👍".
        - Acknowledge the user's intent, e.g., "I understand you, sir!", "Absolutely, let me help with that!", "Great question!".
        - Always include supporting chunk references in parentheses, e.g., (Chunk 123).
	- PROCEDURAL / STEP-BY-STEP QUESTIONS (e.g., “What are the steps to apply for a loan?”):
        - Use a clean, organized bullet-point list.
        - Add 1–2 friendly sentences before or after the list, and feel free to open with a conversational phrase or emoji.
	- END all answers with a short, friendly follow-up prompt or suggestion, e.g.,
        - "Would you like me to also show you an example calculation? 😊"
        - "Anything else I can check for you"
        - "Let me know if you want more details


VERY IMPORTANT:
4) If CONTEXT or CONVERSATION SO FAR lacks sufficient information (for any question type), reply exactly: 
   I apologize, I do not have knowledge about this

5) Never hallucinate:
   - If uncertain or if information is not in the CONTEXT or CONVERSATION SO FAR, fall back to the exact apology text above.
   - Do not add disclaimers, filler, or external world knowledge.

6) Keep style informal, super friendly, and conversational.
   - Start some answers with little exclamations or emojis (e.g., "Ahh! 😊", "Alright! 👍", "Haha, I get you!").
   - Use "I understand you" or similar to acknowledge user queries.
   - Keep text short, positive, and clear.
   - Use bullet points when the user question is procedural.

CONVERSATION CONTEXT RULES:
- If user gives short affirmations ("yes", "yeah", "sure", "okay") immediately after you asked a question, treat it as agreement to that specific question
- Continue the conversation thread naturally based on what you just asked
- Don't lose context or start over

CONTEXT ANALYSIS:
- If previous response ended with "Would you like me to guide you through a specific loan product?" and user says "yes", then proceed to ask about loan types
- If previous response asked about loan type and user says "yes" or gives a loan type, continue with that loan's details

CONVERSATION SO FAR:
{past_turns}

CONTEXT:
{context}

User's question:
{question}

Answer:
"""
    prompt = PromptTemplate(
        input_variables=["past_turns", "context", "question"],
        template=template,
    )

    formatted_prompt = prompt.format(
        past_turns=past_turns_context or "",
        context=chunk_context or "None",
        question=query,
    )

    # --- LLM selection ---
    if model_type == "openai":
        llm = ChatOpenAI(
            model="gpt-3.5-turbo",
            temperature=0.0,              
            api_key=os.getenv("OPENAI_API_KEY"),
        )
    elif model_type == "groq":
        llm = ChatGroq(
            model="meta-llama/llama-4-scout-17b-16e-instruct",
            api_key=os.getenv("GROQ_API_KEY"),
            temperature=0.0,
        )
    else:
        raise ValueError("Invalid model type")

    response = llm.invoke(formatted_prompt)
    answer_text = getattr(response, "content", response)

    usage = getattr(response, "usage_metadata", {})
    input_token_count = usage.get("input_tokens", 0)
    output_token_count = usage.get("output_tokens", 0)
    total_token_count = usage.get("total_tokens", 0)
    cost = compute_llm_cost(input_token_count, output_token_count)

    return {
        "text": answer_text.strip(),
        "input_tokens": input_token_count,
        "output_tokens": output_token_count,
        "total_tokens": total_token_count,
        "cost": cost
    }