requirements.txt
streamlit
lancedb
sentence-transformers
langchain
langchain-groq
pandas
numpy
langchain-core
langchain-community
python-docx
pypdf2
altair
python-dotenv
langchain-openai
ftfy
psutil
scikit-learn

fastapi
uvicorn
pydantic
python-multipart

ragas
datasets
boto3
botocore

pyarrow
s3fs
boto3

langgraph